% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/awesome_geodata_table_html.R
\name{awesome_geodata_table_html}
\alias{awesome_geodata_table_html}
\title{generate html table}
\usage{
awesome_geodata_table_html(table_data)
}
\arguments{
\item{table_data}{a tibble.}
}
\value{
The html table.
}
\description{
generate html table
}
