% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/awesome_geodata_table_read.R
\name{awesome_geodata_table_read}
\alias{awesome_geodata_table_read}
\title{read table}
\usage{
awesome_geodata_table_read(filepath)
}
\arguments{
\item{filepath}{filepath.}
}
\value{
a tibble.
}
\description{
read table
}
